requirements.txt
fastapi
uvicorn
sentence-transformers
feedparser
scikit-learn
python-dotenv
neo4j